requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-tinymce==3.4.0
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2022.2
